+++ RELEASES.md
@@ -5,6 +5,8 @@ Version 0.8.0 (not yet released)
   as `Integer::cmp0`, `Rational::cmp0` and `Float::cmp0`.
 * Move `rug::AssignRound` to `rug::float::AssignRound`.
 * Add `Integer::clamp`, `Rational::clamp` and `Float::clamp` methods.
+* Add `DivRemRounding`, `DivRemRoundingAssign` and
+  `DivRemRoundingFrom` traits, and implement them for `Integer`.
 
 Version 0.7.0 (2017-30-09)
 ==========================